Merge the ignored and pass-through list types.
The only difference between them was that the 'ignored' node list type stopped the
visiting process earlier, i.e. avoids the `super().visit(node)` call. Such a thing is not
a problem if that list is complete, i.e. has all non-interesting nodes (for CSE purposes).

The 'ignored' list is made by inspecting the Python grammar. Each node type reachable by
the `expr` rule is taken into consideration. From those, the ones that are not supposed to
be CSE'd are added to the list.