[FSDP] Avoid printing incorrect warning for _get_param_to_fqns
There exist a hack for `_get_param_to_fqns` and `_apply_to_modules`. The condition for the warning of the hack is incorrect and result in overwhelming message for users. This PR fixes the issue.

The original hack is not removed. It will once the support of DMP + FSDP is deprecated.

Differential Revision: [D43135611](https://our.internmc.facebook.com/intern/diff/D43135611/)

ghstack-source-id: 179756141
Pull Request resolved: https://github.com/pytorch/pytorch/pull/94494